+++ NEWS.md
@@ -1,5 +1,7 @@
 # v0.2.5.9001
 * A new `cl_delete_corpus()`-function can remove loaded corpora from memory.
+* Function `cl_charset_name()` is exposed, it will return the charset of a 
+corpus. Faster than parsing the registry file again and again.
 
 # v0.2.5
 * In Makevars.win, libiconv is explicitly linked, to make RcppCWB compatible with new